[FIX] hr_payroll: regenerate fields when changing structure's country
Problem: Given that the fields for the salary rules has already been generated
for the current country set on the salary structure, an error occurs when accessing
the payroll reporting records after the user changes the country
because the payroll reporting records will reference a different field name.
The outdated fields will not properly unlink, new fields won't be generated
based on the new country set, but the payroll report will reference the new field name.

Purpose: When changing the country on a salary structure, the rules that appears
on the payroll reporting needs to unlink outdated fields and regenerate the fields
based on the new country value. The salary rule fields are dependent on the country
due to its naming convention.

Steps to Reproduce on Runbot:
1. Install Payroll
2. Create a new salary structure with no country
3. Create a new salary rule and enable "View in Payroll reporting"
4. Check fields and observe that a field prefixed by 'x_l10n_xx' was created
5. Change the country set on the salary structure
6. Try to access a payroll report record and a traceback is thrown

opw-4649030

closes odoo/enterprise#87550

X-original-commit: 897a84b9dd722d720804e5c6eea628b4ff9fd1c8
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Mylyna Hy (myhy) <myhy@odoo.com>