[FIX] l10n_de_reports: display correct amount if exporting through datev
Steps to produce :

1) Install 10n_de_reports and switch to the German company
2) Create a journal entry from accounting with a credit and debit
3) Now update the credit and debit to a different value.
4) Post the journal
5) Go to "Accounting / Reporting / Audit Reports / General Ledger"
6) Download "DATEV DATA (ZIP)" and check "EXTF_accounting_entries.csv"

Issue:
The line for the created journal entry has a price_total of the previous balance
Even after updating the credit and debit values.

Cause:
When the user updates the 'debit' and 'credit' values for an `aml`,
The `price_totals` value remains unchanged as it is not dependent on `balance`.

So, it will show the previous value in the report line because we are taking
the abs(aml.price_total) from the report data.

Solution:

We can add an extra check of the move type so that it will fall back
to the else block, where the line_amount will take a value by
calculating taxes from `aml.tax_ids.compute_all`.

opw-4707567

closes odoo/enterprise#86762

X-original-commit: adefa820401ab855934627671b593774b7d262d7
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>