[FIX] l10n_mx_edi: Fixing the complementos injector.
In PR 191823 in the comunity repo, a new functionality was added to allow a user with salesman permissions to send and print the invoice. However this change had unexpected effects in the mexican localization, precisely in the _clean_cfdi_values function which now removes all values from the dictionary that are not plain text.

The complemento's injector was using the addenda value from the dictionary which was a recordset of the l10n_mx_edi.addenda model, with these changes, this recordset was removed from the dictionary, causing the complemento to never be added.

To fix it, the values relevant to the complemento have been marked as unsudo, thus preventing them from being removed in the clean_cfdi_values function.

closes odoo/enterprise#84064

X-original-commit: 63b8be3ff9480b5ce1a58b4aa83ecc00f3bc10b1
Signed-off-by: Laurent Smet (las) <las@odoo.com>