[FIX] helpdesk_timesheet: avoid compute loop
This commit makes sure we will not have a loop between compute of
task_id, ticket_id and project_id since one of them could be altered by
the user which triggers the compute of the other fields. By avoiding the
compute loop, we avoid erasing the choice of the user.

task-4284216

Part-of: odoo/enterprise#81686
Related: odoo/odoo#202313
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>