[FIX] l10n_uk_reports_cis: Add rounding to payment and materials column
Before, the feature was not usable because HMRC would only accept whole pounds for the total payments made and the materials cost.

The fix was to round the values on the report to whoule pounds for those two columns.
This way, it would become valid for HMRC as the return is based on those report values.

opw-4686819

closes odoo/enterprise#85057

X-original-commit: ffb0aa8d7aad913328bb0b585197172da4e838b3
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>