[FIX] account_followup: recipients cc template in followup report
Steps to reproduce:
- Open Follow-up levels for 15 days
-Check the option for “Automatic”
- Go to the email template set in the “Content Template” and set a test email in the CC of the email configuration.
- Open the follow up reports for a sample customer who is in the “15 days” stage and change the next reminder date to be a day before today. The reminder must be set to automatic.
- There will appear a red ribbon on the right corner that says “In need of action”
- In another tab, open scheduled actions and Open the action “Account Report Followup; Execute followup”
- Run this action manually

Issue:
Go to the previous tab and notice that the email is only sent out to the customer partner record and not the cc email address

Note:
This issue only occurs if the automatic followup is set. If we try to manually followup by clicking on the “Follow Up” button on the customer report, the email is sent correctly to both addresses.

Cause:
When running the cron, the recipients aren't in the options

Solution:
Add an helper to find all necessary recipients from the template if no recipient ids in the options

Note fw saas-18.1:
As from https://github.com/odoo/odoo/commit/5e88a2f1b54559219ed9df6ae5aad7c71cb7384f
`use_default_to` is set to True by default by-passing the use of cc etc
https://github.com/odoo/odoo/blob/d6bf6df30095a27eac0315ea78bb57b721efdd5e/addons/mail/models/mail_template.py#L360-L363

opw-4527398

closes odoo/enterprise#88309

X-original-commit: 1185b69579741d58afbac9c8d5184c4b5dc281f0
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Yolann Sabaux (yosa) <yosa@odoo.com>