[FIX] helpdesk: final step of tour should be a check
This error shows up in 18.2, possibly because by waiting more for
browser responses #206271 made it more likely that this error would
have the time to be sent back down the pipe (also might be a change in
chatter impl which makes 18.2 more sensible). However it's likely a
global issue: when the tour "click[s] on ticket", it does a full page
navigation then triggers the async load of a chatter which can take a
pretty long while.

It seems like one of the tour teardown steps can cause the loading of
the chatter bundle to get cancelled, which is then raised as an
exception. The error is triggered from `getBundle` while reading the
response body, but it just wraps and forwards the error it gets, and
swallowing all bundle errors is probably a bad idea...

https://runbot.odoo.com/odoo/error/134748

closes odoo/enterprise#83874

X-original-commit: 1d3666a9153994779d047c711f055eb26f5396f8
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>