[FIX] web_enterprise: invalidate persistent caches when database structure changes.
Before this commit, the persistent caches were not invalidated when the
database structure changed (e.g.: installing/removing modules; adding/removing
fields). The persistent caches will return old values, which could lead to
errors. This can also happen when using a different database with the same URL
(e.g.: in local when developing; or changing the database using
`web/database/manager`; or changing the cli parameter
`-d<database>,--database<database>`).

Now the persistent caches will be invalidated if the database structure is
changed, or if a different database is used.

closes odoo/enterprise#83882

Related: odoo/odoo#206815
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Co-authored-by: Rémy Voet <ryv@odoo.com>