[IMP] delivery_starshipit: reliability improvement
We are collaborating with Starshipit to improve the
quality of the integration; allowing Odoo to be
recognized as a platform and improving the fetching
of exact price to be more reliable.

We also no longer block the order if the exact
price cannot be fetched in time.

task-4748113

X-original-commit: 9018f310ee46b76339a878b8bf4826cae07df3a1
Part-of: odoo/enterprise#89254
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Vincent Sevestre (vise) <vise@odoo.com>