[FIX] stock_barcode_mrp: generate all multistep transfers for barcode MO
Task: 4558792

If 3-step manufacturing is enabled, the creation of an MO should also
create 2 stock transfers: "Pick Components" and "Store Finished
Product". Currently, when an MO is created via Barcode, only SFP is
created.

The problem boils down to the newly created stock move not having a
`warehouse_id` field set, which causes Odoo to miss the required stock
rules. This PR fixes the issue by fetching the `warehouse_id` field in
XML, from where it is later passed to JS and back to Python's `create`
method.

Additionally, by-product moves for MOs created via Barcode were not
correctly associated with the MO. This issue was solved by passing the
missing fields from XML to Python.

closes odoo/enterprise#90679

X-original-commit: afda632b347fa5ce6a2c13a3e8f5160b60117361
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Pawel Fertyk (pafe) <pafe@odoo.com>