[FIX] knowledge: update templates using file block to the new format
This update is done without an upgrade because the old format will still work.
But for cleanliness, templates should be reflect what the editor will actually
do, and since [This commit], files are `span` elements, not `div` anymore, and
they are wrapped in a baseContainer (`div` or `p`).

[This commit]: https://github.com/odoo/odoo/commit/96c8c398c0fbef519b56edf4941691d11372eac0

task-4745902

closes odoo/enterprise#85220

X-original-commit: ef7e746f63502777598695cb8c8d4ed7f78bf834
Related: odoo/odoo#209372
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>