[FIX] pos_settle_due: hide settle button if no payment method
Currently you are able to select the button to settle customer account/
deposit money even though the config does not have access to a paylater
payment method. This will never settle the customer account.

Steps to reproduce:
-------------------
* Remove the customer account payment method from the config
* Open the pos session
* Find a customer that has an amount due
* Select "Settle due accounts"
* Process
> Observation: The customer account is never settled.

Why the fix:
------------
If there is no payment method of type "pay_later" we should not show the
button whose purpose is to use such payment method.

We revert this commit https://github.com/odoo/enterprise/commit/0eae863a9155a188d954b11e14212179953dbb67
as this fixes the problem earlier and we will no longer be
able to reach the fix in the commit mentionned.

opw-4488571

closes odoo/enterprise#86374

X-original-commit: 0ea459d124c9e15597333d16179ce909d93ed649
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>