[PERF] appointment: Speed up loading of the default kanban view
Description:
------------
For non-admin users, loading the default kanban view of the Appointment
application triggers the compute method
`_compute_appointment_counts`. This is quite slow as it calls an
override of `_read_group`, which adds an elaborate domain for
privacy in `_get_default_privacy_domain`.

This patch optimizes domains to generate more efficient queries by:

- Simplifying useless sub-queries of the form `fkey in (select id from comodel
where id = X)` to `fkey in (X)` where it makes sense (in `sudo` context)
- Adding supporting indexes

Benchmark:
----------
On odoo.com, the time for a regular user to open the default kanban view
of appointments is:

| Before (hot) | After (hot) | Speedup |
|--------------|-------------|---------|
| 11s          | 1.2s        | 9.2x    |

Reference:
---------
task-4744275

closes odoo/enterprise#85477

X-original-commit: 9202d8f6458778dbab2c21283754ad59bde2e213
Related: odoo/odoo#209898
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>