[FIX] sale_subscription: amount and amount_total of paymentlink on SO
Before this commit,when a payment link was generated for a recurring SO
already invoiced, the default values of amount and amiunt_max would take
into account all previous transactions. As a result, the amount would be
badly computed as recurring order have many transactions (at least once or each
period). This commit ensure to use the next invoice amount.

closes odoo/enterprise#90546

Taskid: 4352396
X-original-commit: bed7c73fb72fe52eb496aaf9627ec755b0693da2
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>