[IMP] sign: refuse shareable link
Before this commit, when refusing a document as a public user,
it was not possible to identify the user, we only asked for the
refusal reason and that was it. Also, there was no way to refuse
a shared document, only sent ones.

After this commit, we now can refuse shared documents, and they
keep created for allowing other users to refuse them, including
public users. When the public user is refusing it, we ask for
their name and email for identifying them (and creating a partner
from it).

task-4635602

closes odoo/enterprise#83927

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>