[FIX] voip: fix missing getter for contact.
The incoming invitation interface didn't display contact button althought the caller is already in the contacts. Also, after accepting a call from a caller that is in the contact, a button "Add contact" was dispalyed. This happened because the template was trying to retrieve a `contact` variable but there was not getter implemented for this variable.

closes odoo/enterprise#83640

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>