[FIX] appointment: add company to new appointment customer search
**Steps to reproduce:**
- Set up CRM / Appointment / Website apps
- Create new company (2)
- Create partner contact with specific email and the new company
- Create appointment type linked to a staff user of another company (1)
    who has no access to the new one
- Enable `Create Opportunities` on the new appointment type
- Go to the website as an anonymous user
- Submit an appointment request using the partner's email
- `Incompatible companies on records:` error

**Issue:**
Previous solution (17.0) was trying to use existing partner if the email was provided and existed in the database
`customer = request.env['res.partner'].sudo().search([('email_normalized', '=', email_normalized)], limit=1`
This means that, in a multi-company environment, the lead created by the new appointment for the appointment
user of company 1 can be linked to the partner contact of company 2 and throw an error on validation.

**Fix:**
Restricted partner search domain for the customer, using current allowed companies
with `staff_user` or the user which created the `appointement_type`.

Changes were dropped in 18.0 as the partner matching method was updated, but
the test was kept.

opw-4713060

closes odoo/enterprise#89578

X-original-commit: 0de8e7b3cdbf0ef705e11d9f9d1a97257df7cc9c
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>
Signed-off-by: Florentin Delcourt (defl) <defl@odoo.com>