[FIX] l10n_no_saft: saft bugs
This commit solves the following bugs in Norwegian saft as per customer feedback:
1. Removed the StandardAccountID
2. Added the CreditAmount tag value to the XML file if the both debit and credit values is 0 for an entry

The latter point in needed because the SAF-T XML validator expects CreditAmount/DebitAmount tag even if the entry is empty.
This case may happen if, for example, we have an exchange rate entry = $0.000001.
Since the journal entry is approximating to only 2 decimal places, the amounts on the entry end up being 0s.

task-4714764

closes odoo/enterprise#88612

X-original-commit: ce9eed675db4de3bcd1b283688c5c44d2e1ecf12
Signed-off-by: Hesham Saleh (hsal) <hsal@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>