[FIX] spreadsheet_edition: gc only with snapshot
The revision auto-vacuum mistakenly deletes spreadsheet initial attachment
if the spreadsheet has never been snapshotted (but the spreadsheet has not
been updated for more than a year).

In this case, there are active revisions (but no inactive revisions, and no
snapshot). The initial attachment `spreadsheet_data` should not be reset
with `spreadsheet_snapshot` which is `False`

Task-4749537

closes odoo/enterprise#84639

X-original-commit: 98ac6d2a6f0fef13d371ba9d13581eaf8e45c5de
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>