[FIX] sale_subscription: Fix key and field errors in sale order line sale_subscription test
This commit address keyerror `default_pricelist` which is already defined in superclass but not used in the test case.
and
This commit fixes a `Field error` in the `test_postpaid_next_invoice_date` test case in the `sale_subscription` module.
The error was caused by referencing an invalid field name (`tax_id`) on the `sale.order.line` model during the creation of a sale order.
The correct field name, `tax_ids`, has been used instead.

build_error-181599

closes odoo/enterprise#84357

X-original-commit: 498ee61682c79e930e10f60b0a15a2a5dd7fbd09
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Omar Sherif Ali Hassan (osah) <osah@odoo.com>