[IMP] voip: add no-search-results buttons.
In voip, you may search for different things through different tabs. For
example, you may search for recent calls in the recent tab, or for
specific contacts in the contacts tab. When there is not matching
result, there should be a button that guides you to a some page that may
be relevant to your search. For example, if you can't find a specific
contact in the contacts tab, you may have a button to direct you to
create that contact as it seems like you don't have it on your list.

This commit introduces those guide button on the following tabs:
`contacts`, `recent_calls`, and `activities`.

closes odoo/enterprise#84028

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>