[REM] account_sepa_direct_debit: Remove sdd_mandate_id from account.move
The sdd_mandate_id field on account.move was incorrectly set and conceptually unnecessary, as a mandate should not be directly linked to an invoice.
This link was only used for the "Invoices Paid" smart button on the mandate view.
Since account.payment already holds the sdd_mandate_id and invoices have access to related payments, this field is redundant and has been removed.

task-4531618

closes odoo/enterprise#84031

Related: odoo/upgrade#7607
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>