[FIX] social_facebook: add optional business_management scope
This commit fixes an issue with the Facebook app and the management of pages.
For Facebook to identify the user as able to manage a page, we need to provide
another permission to the scope.
The permission `business_management` is needed for the API to understand that
the user is able to manage its pages.

This resolves an issue with companies using their own configuration for the
Facebook pages connection to the social app.
Thus we are adding it to the scope but as an optional permission.

Some users may not have this permission set up for their own applications, thus
we are adding the system parameter `social.facebook_no_business_management`.
This parameter blocks the addition of the permission to the scope if set to any
value, if not set it adds the permission to the scope.

task-4719790

closes odoo/enterprise#86268

X-original-commit: b561bbb2dbcc0c81b730ff09987c662544eceb4c
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>
Co-authored-by: bikh-odoo <bikh@odoo.com>