[FIX] account_reports: fix end balance of Trial Balance with analytic
Steps to reproduce:
- Install Accounting
- Activate "Analytic Accounting" in Accounting settings
- Go to "Accounting / Accounting / Journal Entries"
- Create an entry with some analytic distribution
- Go to "Accounting / Reporting / Audit Reports / Trial Balance"
- Activate "Analytic Group By" in the settings of the report
- Add the account used in the analytic distribution in "Analytic" group option

Issue:
In the "End Balance" column, all the amounts are summed in the total
column instead of being grouped by analytic accounts.
It results on a total amount being twice what it should be as the
total column of each period is also included in the sum.

opw-4648009

closes odoo/enterprise#84936

X-original-commit: 76df7b314d87ef48584f79eb6b2c5e51fcc08cdf
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>