[IMP] voip{,_hr_applicant}: filter out forbidden records from activities
If an activity is created for a record that the responsible user does
not have access to, trying to fetch "today's activities" would result in
an access error.

This commit adapts the get_today_call_activities code to filter out
activities associated with a record that the current user does not have
access to.

closes odoo/enterprise#84039

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>