[FIX] sale_subscription: fix computation of parent_line_id
Before this commit _compute_parent_line_id make the assumption that
a parent line can only be linked to once child: When the parent_line_id
is set, it's remove from the pool of possible parent_line_id

Actually this is not the case, different upsell order can have a line
that have the same parent.

Upsell order A can have line 1 with the same parent as the line 2 of
Upsell order B.

Before this commit if the parent line of both order were computed at the
same time one of them would end up without parent.

This situation is getting worse since odoo@42008de (saas-18.1)
were all the parent_line_id are recomputed on every upsell order (even
already confirmed one) as soon as line on the parent order are touched

The follow use case lead to duplicate line on the parent

A parent subscription A with 1 product Z
Create an upsell B with a line that upsell 1 product Z
Create an upsell C with a line that upsell 1 product Z
On both upsell the parent line is set properly

Confirm Upsell B, the parent line on Upsell C is set to false
Confirm Upsell B, the parent subscription ends up with a two line for
the product Z

After this commit,
parent_line_id are computed per order, so one order cannot have two
lines with the same parent_line_id but two (or more) different order can have each
one line with the same parent_line_id even if their parent_line_id are
computed at the same time.

closes odoo/enterprise#84451

X-original-commit: d5bd0de6e189008740bc3c5ff6945f54c40cadc2
Signed-off-by: Thibault Francois <tfr@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Co-authored-by: Thibault Francois <tfr@odoo.com>