[FIX] timesheet_grid: prevent double time in grid view
A mismatch between backend and frontend caused the time displayed in the grid to be incorrect.
The cell value was doubled due to a frontend-only calculation, but the actual data was correct on the backend.
Refreshing the page fixed the display issue.

Steps to reproduce:
-------------------
* Open the view form of a timesheet cell (🔍)
* Start the timer in the Hours Spent column
* Go back to My Timesheets
* Stop the timer

> Observation:
timer displayed was previous_timer * 2 + new_timer

Why the fix:
------------
'stopTimer()' waits for the orm call 'action_timer_stop' to return a value that will be added to the current cell value.

opw-4701396

closes odoo/enterprise#84964

X-original-commit: 44e046f585cfc177d893b1da7fb6f9f40c46a87a
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Guillaume Vardakis (gvar) <gvar@odoo.com>