[FIX] voip: fix viewing incorrect caller in VoIP.
This is a fix for a bug that happens from time to time. It was reported
more than once that another user was viewed as a caller instead of
the actual caller.

This happened because when someone calls from a local number, they have
a `0` at the beginning of their number. For that case, we use a wildcard
that matches anything that the phone number is a suffix to. So, it
happened that it matched more than one number when the calling local
number is a suffix to more than one international number. Then, the code
just picks the first one randomly, which may be incorrect.

This commit fixes this bug by removing the logic of using a wildcard when
the call is made through a local number. Instead, we use an exact match.

Task-4707543

closes odoo/enterprise#84367

X-original-commit: d683e53b6f158dcda002317e9b60fb305646a7d6
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>
Signed-off-by: Mohammed Basioni (basm) <basm@odoo.com>