[IMP] account_accountant: limit models visible
In the bank reco widget, applicable models are visible for statement
lines.
We should limit the ones visible to ones that are manual and which
have counterparts that can actually work to produce an entry, so with
an account (and a label that can be matched).
Also filter on the company.

task-4749345

closes odoo/enterprise#85048

X-original-commit: 44e1200b42015d21b80e9fdb8243c6af1f63d5e1
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>