[FIX] l10n_br_avatax: allow portal users to view products on website
Issue:
If `l10n_br_avatax` is installed, portal users are not able to view
any products via the shop on a Website. An inherited field's domain
requires access to the `city_id` from the Website's company's
`res.partner` record, but Portal users cannot access this.

This is only an issue on 18.2+, as this value was originally stored in
the field cache due to a separate workflow in previous versions. Previously,
opening a product would trigger `_compute_fiscal_position_id`, which
would eventually call `_compute_address`. `_compute_address` contains
a sudo, allowing the retrieval of the address fields (like `city_id`)
and would store them in cache. The domain would be able to retrieve from
the cache rather than needing to fetch the values.

In 18.2, the `fiscal_position` workflow was refactored, and the address fields
are no longer stored in cache for the domain `_l10n_br_property_service_code_origin_id_domain`
to access prior to hitting any security limits.

Solution:
We should be safe to `sudo` the `city_id`, allowing the domain to complete.

opw-4732256

closes odoo/enterprise#84818

X-original-commit: 63f12d4a5aeb3421895ebbe25697466759d180ad
Signed-off-by: Laurent Smet (las) <las@odoo.com>