[FIX] l10n_in_hr_payroll: dependency on demo data
`hr.dep_rd` is demo data, it's not available anymore during tests in
the default configuration. So replace it by creating a department on
the fly. And remove the resolution of `employee_al` which is also demo
data *and* seems unused.

While at it, nuke a bunch of instance variables which don't seem
used outside of `setUp`.

https://runbot.odoo.com/odoo/error/161648

closes odoo/enterprise#84350

X-original-commit: ea35d11ea995ffe68f0e0dca13a37d1e121c404b
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>