[IMP] account_accountant{_batch_payment}: Bank Reconciliation widget chatter
The aim of this commit is slightly changing the chatter's behavior in the
bank reconciliation widget.
Before this commit, the user activated/deactivated the chatter by clicking
on a chatter icon available on each statement line. Although this feature is
still existing to activate the chatter, we add a new button on the top corner
right where the user can activate / deactivate the chatter.
We also change the selected move id when you select another statement line,
it's not necessary to click on the chatter icon to do it.
This commit also introduces a cleaner way of working with several components
to activate / deactivate the chatter by using a customer service.

task-4750009

X-original-commit: ca520abe552f5e46a4c9433edb45f8627cd40a51
Part-of: odoo/enterprise#84554
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>