[FIX] ai: correcting permission issues
This commit corrects some permission issues for non-admin users.
Specifically, it allows non-admin users to delete their chat channel
with the agent upon closing the chat window, and corrects the retrieval
of API keys in the `llm_api_service` class.

It also corrects the demo data to have a proper link between the
attachment and the agent, and hides the progress banner from the agent
form for non-admin users. These changes solve permission issues for
chatting with the agent, and accessing its form respectively.

closes odoo/enterprise#84330

Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>