[FIX] account_followup: payable accounts in total due
Steps:

- Create a partner P
- Create an invoice for 200 and a bill for 100
  with P as partner
- Go to P form view
-> The 'Customer Statement' smart button show 200 instead
   of 100

This is because the total due does not take into account
the accounts of type `liability_payable` in the unreconciled aml domain.

With this commit, we adapt the domain to include the good accounts.
We also include deprecated accounts. Although it is unlikely to have
unreconciled aml from these accounts, it is reflected in the customer
statement, so it needs to be computed in the total due.

opw-4664402

closes odoo/enterprise#84574

X-original-commit: c7e32cdc42d239c5425c56b0deed341d5a0bf9c7
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>