[FIX] l10n_uk_reports: prevent KeyError on missing 'client_data' in HMRC Send Wizard
The default_get method of the 'l10n_uk.hmrc.send.wizard' model assumed
the presence of 'client_data' in the context, leading to a KeyError
during test_display_name_new_record.

This fix checks for the presence of 'client_data' before attempting to
call import_vat_obligations. If 'client_data' is not present, it avoids
calling import_vat_obligations and raises a UserError, thus preventing the KeyError.

build_error-115767

closes odoo/enterprise#88467

X-original-commit: 621929db25b11eceb228da5a2079ffed4ee0c685
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Omar Sherif Ali Hassan (osah) <osah@odoo.com>