[FIX] ai: prevent ai_tool creation inside search more of ai_topic form
When attempting to create an AI tool from the 'Search More' view within the ai.topic form view, a traceback occurs.
This change resolves the issue by restricting the creation of ai.tool from that view, as ai.tool is not intended to
be generated on the fly by users.

closes odoo/enterprise#84319

Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>