[FIX] l10n_es_reports: fix mod347 BOE cash payments
Fix a bug in mod347 BOE export.

Steps to reproduce:
1- Create an invoice with a spain Company, with an amount lower than 3 005,06€
2- Add a Type for mod347
3- Create a cash payment
4- Export the mod347 BOE

The partner will appear in the BOE with all line at 0. But this partner
shouldn't be in the export.

This is due because of a search on account.partial.reconcile, which
add partners to the export if a cash payment is found in the period.

But this search is not usefully as there is no legal indication that these
partners should be in the export in this case, as the partners should only be
returned by the main queries.

task-4613012

closes odoo/enterprise#85685

X-original-commit: 5dac81e36ede0142a854e0a821a329bfed47a927
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Igor Bertrand (igbe) <igbe@odoo.com>