[REM] mrp_mps: remove Batch Size from MPS Product Configuration
Removed the Batch Size setting from the MPS product configuration.
With the introduction of the standard batch size on the BoM, this setting has
become redundant.
Batch size is now defined and managed directly through the BoM.

As a result, the replenishment quantity in the MPS will always be a multiple of
the BoM's batch size.
This ensures that orders created from the MPS are split according to the defined
batch size, making it easier and more consistent for users to manage production
quantities.

This change promotes the use of a standardised batch size for manufacturing
operations and improves overall usability and convenience.

Task ID: 4395548

Community PR: [208275](https://github.com/odoo/odoo/pull/208275)

closes odoo/enterprise#84605

Related: odoo/odoo#208275
Related: odoo/upgrade#7638
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>