[FIX] voip: fix failing DND selector test
"Do not disturb selector changes state correctly with limited time" is
failing with this error:

```
9. [toBe] Failed to find 1 of "p" with text "Until Jan 1, 2025, 1:15 AM" (Timeout of 3 seconds). Found 0 instead.
> Expected: true
> Received: false
```

Depending on the environment, it seems that the time is sometimes
formatted with a nonbreaking space before the AM, and sometimes not.
This is likely the reason for the failure.

This commit adapts the test in order to exclude the problematic space
from the selector.

closes odoo/enterprise#84433

X-original-commit: d19c3a750f912bf8fb10a91f968c9634b94f33fa
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>