[FIX] account_accountant,account_online_synchronization: Fix traceback
The aim of this commit is fixing a traceback when we have a statement
group with no balance passed as props. It could happen when we have
statement lines without any statement between 2 groups of statement
lines with statement.
This commit changes the component to set this props as optional (as
before) and adds a new props "isValid" to put the label of the
statement group in text-danger when the statement is not valid.

no task id

closes odoo/enterprise#84447

X-original-commit: 1b8f142f375f2e46693c61bf81db7b556db072a3
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>