[FIX] documents, documents_spreadshet: Allow users to use studio on documents
Steps:
- install `web_studio` and `documents_spreadsheet`
- open documents
- open studio on the spreadsheet kanban view (the default one)
- change sort by field to "created on" field
- error

This commit replaces encodeURIComponent with window.encodeURIComponent,
because owl won't try to evaluate this variable via the context.

And so the fix

```js
    get renderingContext() {
        const context = super.renderingContext;
        context.encodeURIComponent = encodeURIComponent;
        ...
    }
```
In `DocumentsKanbanRecord` is no longer necessary.

The error occurred because `encodeURIComponent` was not found in the
context object.
The reason this fix doesn't work with studio is the view is defined as
```xml
<kanban js_class="documents_kanban"
```
and studio does not load view js classes.

And since the fix is in `documents_kanban`, it's not taken into account.
opw-4744886

closes odoo/enterprise#85361

X-original-commit: b96c48317fbbe6f1193d138cd57a2fc8379969b5
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Achraf Ben Azzouz (abz) <abz@odoo.com>